requirements.txt
streamlit
langchain
langchain-community
langchain_groq
sentence-transformers
faiss-cpu
PyMuPDF
python-docx
beautifulsoup4
requests
tavily-python
google-generativeai
